Changed int to long

+++ src/Infrastructure/Presistance/Services/IRedirectQuery.cs
@@ -5,7 +5,7 @@ namespace Presistance.Services
 {
     public interface IRedirectQuery
     {
-        public Task<Redirect> Find(int id);
+        public Task<Redirect> Find(long id);
         public Task<Redirect> Find(string url);
     }
 }
